+++ src/main.cpp
@@ -21,6 +21,7 @@
 #include "power/PowerHAL.h"
 
 #include "FSCommon.h"
+#include "Power.h"
 #include "RTC.h"
 #include "SPILock.h"
 #include "Throttle.h"
@@ -28,7 +29,6 @@
 #include "concurrency/Periodic.h"
 #include "detect/ScanI2C.h"
 #include "error.h"
-#include "power.h"
 
 #if !MESHTASTIC_EXCLUDE_I2C
 #include "detect/ScanI2CConsumer.h"
