+++ src/mesh/http/ContentHandler.cpp
@@ -9,8 +9,8 @@
 #if HAS_WIFI
 #include "mesh/wifi/WiFiAPClient.h"
 #endif
+#include "Power.h"
 #include "SPILock.h"
-#include "power.h"
 #include <FSCommon.h>
 #include <HTTPBodyParser.hpp>
 #include <HTTPMultipartBodyParser.hpp>
