+++ src/modules/Telemetry/EnvironmentTelemetry.cpp
@@ -7,6 +7,7 @@
 #include "EnvironmentTelemetry.h"
 #include "MeshService.h"
 #include "NodeDB.h"
+#include "Power.h"
 #include "PowerFSM.h"
 #include "RTC.h"
 #include "Router.h"
@@ -17,7 +18,6 @@
 #include "graphics/images.h"
 #include "main.h"
 #include "modules/ExternalNotificationModule.h"
-#include "power.h"
 #include "sleep.h"
 #include "target_specific.h"
 #include <OLEDDisplay.h>
