+++ variants/nrf52840/ELECROW-ThinkNode-M6/variant.cpp
@@ -19,8 +19,8 @@
 */
 
 #include "variant.h"
+#include "Power.h"
 #include "nrf.h"
-#include "power.h"
 #include "wiring_constants.h"
 #include "wiring_digital.h"
